refactor(page): split into server shell + client HomePage component (U4)
src/app/page.tsx is now a thin server component that reads
RESEND_API_KEY (and DEMO_MODE) and forwards an emailEnabled boolean to
the new <HomePage /> client component in src/components/home-page.tsx.

Behavior is unchanged for U4 — the prop is unused. U5 wires it into
the email button visibility gate. Tests render <HomePage emailEnabled
={false} /> directly.
